rewardsweb/requirements/development.txt
-r base.txt
## cache requirements
django-redis>=5.4.0
redis>=5.2.1
## tests requirements
pytest>=8.3.5
pytest-cov>=6.0.0
pytest-django>=4.10.0
pytest-mock>=3.14.0
pytest-xdist>=3.6.1
# profiling requirements
django-debug-toolbar>=5.1.0
django-debug-toolbar-template-profiler>=2.1.0
# linting
black>=25.1.0
pylint>=3.3.6
isort>=6.0.1
# docs
Sphinx>=8.2.3
sphinx-rtd-theme>=3.0.2
rope>=1.13.0
ruff>=0.11.2
# ipython
ipython>=9.0.2
